connector/requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.
MarkupSafe==2.1.3
Werkzeug==2.2.3
gunicorn
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
backoff==2.2.1
blinker==1.7.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cohere==4.34
cryptography==41.0.5
dataclasses-json==0.6.2
exceptiongroup==1.1.3
fastavro==1.8.2
frozenlist==1.4.0
greenlet==3.0.1
idna==3.4
importlib-metadata==6.8.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.334
langsmith==0.0.63
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.1
packaging==23.2
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tenacity==8.2.3
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
validators==0.22.0
weaviate-client==3.25.3
yarl==1.9.2
zipp==3.17.0
